+++ source/editor/messagebox.h
@@ -20,9 +20,9 @@ MessageBox.h			: quick helper function to call up a message box
 CLASS DESCRIPTION
 MessageBox:
 **************************************************************************************************/
-extern ULONG gameResourceHandle;
+extern uint32_t gameResourceHandle;
 
-inline int EMessageBox(int MessageID, int CaptionID,ULONG dwS )
+inline int32_t EMessageBox(int32_t MessageID, int32_t CaptionID,uint32_t dwS )
 {
 	char buffer[512];
 	char bufferCaption[512];
